backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.1
pytz==2023.3
requests==2.31.0
django-extensions==3.2.3
pydantic==2.5.3
